Renamed solution to `Stellar` and project `Engine` to `StelarEngine`
