+++ StellarEngine/source/engine/application.cpp
