+++ StellarEngine/source/engine/components/camera/camera.h
@@ -1,0 +1,16 @@
+#pragma once
+#include <glm.hpp>
+#include "../../core.h"
+
+class ENGINE_API Camera
+{
+public:
+	glm::vec3 cameraPos   = glm::vec3(0.0f, 0.0f,  3.0f);
+	glm::vec3 cameraFront = glm::vec3(0.0f, 0.0f, -1.0f);
+	glm::vec3 cameraUp    = glm::vec3(0.0f, 1.0f,  0.0f);
+
+public:
+	void SetCameraPosition(float x, float y, float z);
+	void SetCameraFront(float x, float y, float z);
+	void SetCameraUp(float x, float y, float z);
+};
