+++ StellarEngine/source/engine/core.h
