+++ StellarEngine/source/engine/entry.h
