+++ StellarEngine/source/engine/rendering/opengl_pipeline.cpp
